requirements.txt
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.1.0
Markdown==3.4.1
pytz==2022.6
sqlparse==0.4.3
zipp==3.11.0
asgiref==3.5.2
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.1.0
Markdown==3.4.1
Pillow==9.3.0
Pygments==2.13.0
pyngrok==5.2.1
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
zipp==3.11.0
